GH-14 Externalize logic of grouping records into files
This commit introduces `RecordGrouper` interface and an implementation
`TopicPartitionRecordGrouper` that encapsulates the logic of grouping
`SinkRecord`s into files by topic and partition currently implemented in
`GcsSinkTask`. Also, the commit adds some tests for it.

The purpose of this change is to enable adding new grouping logics like
by records' key.