Replace YYYY with yyyy in template formatting
Originally we introduced `{{timestamp:unit=YYYY}}` to render the year of the timestamp in the template. We passed YYYY directly into Java's timestamp formatting. However, `YYYY` is bit different from `yyyy`, which users actually want.

Unfortunately, the connector was release and use with `YYYY` for some time, so backward compatibility is important.

This commit makes `YYYY` -> `yyyy` substitution in the configuration before it comes to validation, etc., i.e. at the earliest possible. This allows to keep backward compatibility. On the other hand, `YYYY` is stripped from `README` and help messages.