Enhance GcsSinkConfigTest with ConfigDef.validate
This commit makes `GcsSinkConfigTest` to also check that `ConfigDef.validate` passes (or fails in expected places) apart from the same checks done inside it after its constructor.

When `GcsSinkConfig.handleDeprecatedYyyyUppercase` was added, it made `correctlySupportDeprecatedYyyyUppercase` test pass. However, when the connector is created for real, its config is validated as is by `ConfigDef.validate`, without any modifications done further in `GcsSinkConfig`. If we check this, `correctlySupportDeprecatedYyyyUppercase` fails. To cover this (and similar) cases, this commit increases the complexity of the tests.