feat: Add Avro Container File output support
The Avro Object Container File support for GCS connector.

The integration tests are refactored to use single shared
Kafka container. This is to reduce the number of created networks
for containers. The testcontainers do not prune the networks when
container is stopped, the container is now started and stopped once
for each test class instead of once per each test.