Implement Ring Buffer for S3 API (#397)
The S3 Api has a possible race condition in some advanced scenarios.
In the diagram below we can see that the last seen file that was
processed is of a higher lexigraphical order then one file which was
larger and was not available on S3 during the original query. In the
current implementation this file is never read or processed.
<img width="1538" alt="{9B08ED12-AF10-40AE-B60C-F717F2C2D9EC}"
src="https://github.com/user-attachments/assets/ca3f4036-b00a-4577-8281-1fdd988e9534"
/>

Having looked at it I have added a ring buffer which by default acts as
the current implementation does but if wanted by users can be made as
large as they need to ensure that slow to upload files are never missed.
All the time, because we still have the processed filename in the ring
buffer, we do not need to reprocess them as we can tell they have
already been processed.

Lastly a bug where removeEntry was called on commitRecord has been
removed, and replaced with removeEntry on the Object being fully
processed. commitRecord is called when the message has been sent and
acked by Kafka or a Transformer not when the offsets have been saved to
the offset topic.

Signed-off-by: Aindriu Lavelle <aindriu.lavelle@aiven.io>