Aindriu aiven/azure blob client (#418)
PR #414 should be closed first.

This PR introduces the Azure Config Fragment which handles all the Azure
specific configuration required by the Source Connector, it also handles
creation of the BlobServiceAsyncClient.

The PR also introduces the AzureBlobClient which handles listing all the
blobs in a container, and handles downloading the of an individual blob.

This is done using an async client, this PR also includes the first
Integration tests, by testing the client against the docker container.

This PR **does not** include changing the Azure sink connector to use
the AzureBlobConfigFragment

Signed-off-by: Aindriu Lavelle <aindriu.lavelle@aiven.io>