Update to readme document to show how to process any files with a regex (#551)
In 3.4.0 we updated the validation to allow the source connector to use
a simple regex which would scan for files to read. This was not
documented and adds a considerable amount of flexibility.


Also updated the fact that the sink connector compression defaults to
gzip and the source defaults to none.

---------

Signed-off-by: Aindriu Lavelle <aindriu.lavelle@aiven.io>